+++ src/__init__.py
@@ -1,3 +1,0 @@
-__version__ = '3.0.1'
-
-APPLICATION_NAME = 'w2re'