+++ tests/unit/test_command_line_w2re.py
@@ -10,17 +10,18 @@ from unittest.mock import (
     patch,
 )
 
-from src.formaters import (
+from w2re.command_line_w2re import main
+from w2re.formaters import (
     ALL_FORMATERS,
     PythonFormater,
 )
-from src.w2re import main
 
 
 class Main(TestCase):
     def setUp(self):
-        mock_stream_to_regexp_patcher = patch('src.w2re.stream_to_regexp')
+        mock_stream_to_regexp_patcher = patch('w2re.command_line_w2re.stream_to_regexp')
         self._mock_stream_to_regexp = mock_stream_to_regexp_patcher.start()
+        self._mock_stream_to_regexp.return_value = ''
         self.addCleanup(mock_stream_to_regexp_patcher.stop)
 
     @patch('sys.stdout', new_callable=StringIO)
@@ -44,8 +45,9 @@ class Main(TestCase):
             self._mock_stream_to_regexp.reset()
 
     def test_it_refuses_unknown_formaters(self):
-        with self.assertRaises(SystemExit) as exception_context:
-            main(['-f', 'sjaflksag'])
+        with patch('sys.stderr', new_callable=StringIO):
+            with self.assertRaises(SystemExit) as exception_context:
+                main(['-f', 'sjaflksag'])
 
         self.assertNotEqual(0, exception_context.exception.code)
         self._mock_stream_to_regexp.assert_not_called()