+++ w2re/__init__.py
@@ -1,0 +1,14 @@
+from w2re.formaters import (
+    BaseFormater,
+    PythonFormater,
+    PythonWordMatchFormater,
+)
+from w2re.prefix_tree.tree import PrefixTree
+from w2re.utils import (
+    iterable_to_regexp,
+    stream_to_regexp,
+)
+
+__version__ = '3.0.1'
+
+APPLICATION_NAME = 'w2re'