+++ w2re/prefix_tree/primitives.py
@@ -1,6 +1,6 @@
 import re
 
-from src.prefix_tree.letter_range_utils import collapse_letter_ranges
+from w2re.prefix_tree.letter_range_utils import collapse_letter_ranges
 
 
 def get_stop_index(word, sliding_window_length):