+++ w2re/__init__.py
@@ -12,3 +12,5 @@ from w2re.utils import (
 __version__ = '3.0.1'
 
 APPLICATION_NAME = 'w2re'
+CHANGELOG_URL = 'https://github.com/radeklat/words-to-regular-expression/' \
+                'blob/develop/CHANGELOG.md'