+++ w2re/__init__.py
@@ -9,7 +9,7 @@ from w2re.utils import (
     stream_to_regexp,
 )
 
-__version__ = '3.0.1'
+__version__ = '3.1.0'
 
 APPLICATION_NAME = 'w2re'
 CHANGELOG_URL = 'https://github.com/radeklat/words-to-regular-expression/' \